refactor: extract EP collective time validation
